fix: rename kani proofs module to avoid shadowing the kani crate
The proofs module was named `kani`, which shadowed the external kani crate: kani's assertion codegen emits `kani::assert(...)` for instrumented `debug_assert!`s, and that path resolved to the local module (no such fn) instead of the crate, failing every harness with E0425. Rename to `proofs`. Also register `cfg(kani)` in [lints.rust] check-cfg so clippy/doc/x-plat stop failing on -D warnings.
