+++ compact_str/Cargo.toml
@@ -82,3 +82,6 @@ zeroize = { version = "1", default-features = false, features = ["alloc"] }
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(kani)'] }
